feat: 添加Cesium

+++ houtu-math/src/epsilon.rs
@@ -48,4 +48,4 @@ pub const LUNAR_RADIUS: f32 = 1737400.0;
 
 pub const SIXTY_FOUR_KILOBYTES: u32 = 64 * 1024;
 
-pub const FOUR_GIGABYTES: u32 = 4 * 1024 * 1024 * 1024;
+pub const FOUR_GIGABYTES: u64 = 4 * 1024 * 1024 * 1024;
